+++ server/src/main.rs
@@ -1,6 +1,7 @@
 mod client;
 mod fastchat;
 mod websocket;
+mod routes;
 
 #[tokio::main]
 async fn main() {
