+++ server/src/routes/mod.rs
@@ -1,0 +1,4 @@
+mod add;
+
+pub use add::print_client;
+pub use add::{add_handler as add_handler};
